上传课件部分数据库操作

+++ web/course/CourseList.js
@@ -71,8 +71,8 @@ function SearchCourse(){
                     else {
                         var status = "关闭";
                     }
-                    var operation = "<a id='modify' name='"+data.CourseArray[i].coID+"'>修改 </a><a id='delete' name='"+data.CourseArray[i].coID+"_"+data.CourseArray[i].coYear+"'> 删除</a>";
-                    var row = '<tr id="row" name="'+data.CourseArray[i].coID+'"><td>'+data.CourseArray[i].coBH+'</td><td>'+data.CourseArray[i].coName+'</td><td id="coYear">'+data.CourseArray[i].coYear+'</td><td>'+data.CourseArray[i].coTerm+'</td><td>'+data.CourseArray[i].coDept+'</td><td>'+status+'</td><td>'+operation+'</td></tr>';
+                    var operation = "<a id='modify' name='"+data.CourseArray[i].coID+"'>修改 </a><a id='delete"+i+"' name='"+data.CourseArray[i].coID+"_"+data.CourseArray[i].coYear+"'> 删除</a>";
+                    var row = '<tr id="row" name="'+data.CourseArray[i].coID+'"><td name="'+data.CourseArray[i].coID+'">'+data.CourseArray[i].coBH+'</td><td name="'+data.CourseArray[i].coID+'">'+data.CourseArray[i].coName+'</td><td id="coYear" name="'+data.CourseArray[i].coID+'">'+data.CourseArray[i].coYear+'</td><td name="'+data.CourseArray[i].coID+'">'+data.CourseArray[i].coTerm+'</td><td name="'+data.CourseArray[i].coID+'">'+data.CourseArray[i].coDept+'</td><td name="'+data.CourseArray[i].coID+'">'+status+'</td><td id="op">'+operation+'</td></tr>';
                     $("#tablebody").append(row);
                     $("#tablebody").delegate("#row","mouseover",function(){
 						$(this).css({
@@ -84,21 +84,28 @@ function SearchCourse(){
 							cursor : "pointer"
 						});
 					});
-                    $("#tablebody").delegate("#delete","click",function(){
+                    $("#tablebody").delegate("td[id!='op']","click",function(){
+						var coID = $(this).attr("name");
+                        window.location.href = "../courseconfig/Lecture.php?coID="+coID+"";
+					});
+                    $("#tablebody").delegate("#delete"+i+"","click",function(){
 						var str = $(this).attr("name");
                         var array = str.split('_'); 
                         var coID = array[0];
-                        var coYear = array[1];       
-                        DeleteCourse( coID, coYear );
+                        var coYear = array[1];
+                        if(!confirm("删除课程将删除该课程下所有的资料，请确认！")){
+                            return;
+                        }
+                        else {
+                            DeleteCourse( coID, coYear );
+                        }
+                        
 					});
                     $("#tablebody").delegate("#modify","click",function(){
 						var coID = $(this).attr("name");
                         window.location.href = "CourseModify.php?coID="+coID+"";
 					});
-                    $("#tablebody").delegate("#row","click",function(){
-						var coID = $(this).attr("name");
-                        //alert(coID);
-					});
+					
                 }
                 
             }
